recognition/requirements.txt
onnx==1.14.0
einops==0.6.1

# only for mac
onnxruntime==1.16.0; sys_platform == 'darwin'
# only linux or windows
onnxruntime-openvino==1.16; sys_platform != 'darwin'
# Windows-specific dependency
openvino==2023.1; sys_platform == 'win32'

omegaconf==2.3.0
opencv-python==4.8.0.74
tqdm==4.65.0

flask~=3.0.0
python-socketio==5.9.0